requirements.txt
black==22.1.0
flake8==4.0.1
isort==5.10.1
django-widget-tweaks==1.4.12
django-pure-pagination==0.3.0
django-taggit==2.1.0
django-ckeditor==6.2.0
asgiref==3.5.0
click==8.0.4
Django==4.0.3
django-countries==7.2.1
django-js-asset==2.0.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
